btc-dashboard-backend/requirements.txt
requests
pandas
python-binance
ccxt
pydantic
boto3
feedparser
praw
fredapi
yfinance
flask
flask-cors
python-dotenv
schedule

